refactor(app): relocate and update react-router-express
Moved react-router-express to adapters directory for better structure
. Updated import paths accordingly to reflect the new location. This
change improves code organization and aligns with module conventions.
